Moved the django ninja modifications into other file
Did not work if loaded before wsgi by gunicorn
