+++ arkad/event_booking/schemas.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 from uuid import UUID
-from arkad import Schema
+from arkad.customized_django_ninja import Schema
 
 
 class EventSchema(Schema):
