+++ arkad/student_sessions/schema.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 from typing import List, Optional
-from arkad import Schema
+from arkad.customized_django_ninja import Schema
 from user_models.schema import ProfileSchema
