+++ arkad/user_models/api.py
@@ -10,7 +10,7 @@ from django.db import IntegrityError
 from django.http import HttpRequest
 from ninja import File, UploadedFile, PatchDict
 
-from arkad import Router
+from arkad.customized_django_ninja import Router
 from arkad.email_utils import send_mail
 from arkad.jwt_utils import jwt_encode, jwt_decode
 from arkad.settings import SECRET_KEY
