+++ requirements.txt
@@ -12,7 +12,7 @@ constantly==23.10.4
 cryptography==44.0.1
 Django==5.2
 django-cors-headers==4.7.0
-django-ninja==1.3.0
+django-ninja==1.4.1
 django-stubs==5.1.3
 django-stubs-ext==5.1.3
 dnspython==2.7.0
